Added  user service getUserReports method and test for this method

+++ Homework/src/main/java/ignitis_homework/dto/UserReportResponse.java
@@ -11,6 +11,6 @@ public class UserReportResponse {
     private Integer messagesCount;
     private String firstMessageDate;
     private String lastMessageDate;
-    private Integer averageMessageLength;
+    private Double averageMessageLength;
     private String lastMessageText;
 }
